Add config option via environment variables
Before, the app configuration was required to be defined in a
`config.py` file in the Flask "instance" folder. This approach does not
mix well with a containerized app.

To avoid having to copy a configuration file to a build container, the
approach was changed to allow configuration via environment variables.