Configure SECRET_KEY from envvar
Before, the default value for SECRET_KEY was defined in the code.

This put the configuration in two separate places and also allowed for
it passing silently when the key is not defined.

Now, the SECRET_KEY needs to be defined as an envvar. The existence is
also checked and enforced. If the envvar does no exist, the app raises
an exception and does not start up.